+++ ICD.Connect.Conferencing.Cisco/Devices/Codec/Components/Cameras/FarCamera.cs
@@ -1,4 +1,6 @@
-﻿using ICD.Common.Utils.Services.Logging;
+﻿using System.Collections.Generic;
+using ICD.Common.Utils.Services.Logging;
+using ICD.Connect.API.Commands;
 using ICD.Connect.Cameras;
 using ICD.Connect.Conferencing.Cameras;
 
@@ -9,6 +11,28 @@ namespace ICD.Connect.Conferencing.Cisco.Devices.Codec.Components.Cameras
 	/// </summary>
 	public sealed class FarCamera : AbstractCiscoCamera, IRemoteCamera
 	{
+		/// <summary>
+		/// Mapping of eCameraPanTiltAction to cisco command text
+		/// </summary>
+		private static readonly Dictionary<eCameraPanTiltAction, string> s_PanTiltActionToCisco = new Dictionary
+			<eCameraPanTiltAction, string>
+		{
+			{eCameraPanTiltAction.Up, "Up"},
+			{eCameraPanTiltAction.Down, "Down"},
+			{eCameraPanTiltAction.Left, "Left"},
+			{eCameraPanTiltAction.Right, "Right"}
+		};
+
+		/// <summary>
+		/// Mapping of eCameraZoomAction to command text
+		/// </summary>
+		private static readonly Dictionary<eCameraZoomAction, string> s_ZoomActionToCisco = new Dictionary
+			<eCameraZoomAction, string>
+		{
+			{eCameraZoomAction.ZoomIn, "ZoomIn"},
+			{eCameraZoomAction.ZoomOut, "ZoomOut"}
+		};
+		
 		/// <summary>
 		/// The CallId for the remote camera.
 		/// </summary>
@@ -46,7 +70,13 @@ namespace ICD.Connect.Conferencing.Cisco.Devices.Codec.Components.Cameras
 		/// <param name="action"></param>
 		public override void PanTilt(eCameraPanTiltAction action)
 		{
-			Codec.SendCommand("xCommand Call FarEndControl Camera Move CallId: {0} Value: {1}", m_CallId, action);
+			if (action == eCameraPanTiltAction.Stop)
+			{
+				StopPanTilt();
+				return;
+			}
+
+			SendMoveCommand(s_PanTiltActionToCisco[action]);
 			Codec.Log(eSeverity.Informational, "Moving Far End Camera CallId: {0}, Direction: {1}", m_CallId, action);
 		}
 
@@ -55,8 +85,64 @@ namespace ICD.Connect.Conferencing.Cisco.Devices.Codec.Components.Cameras
 		/// </summary>
 		public override void StopPanTilt()
 		{
+			StopMove();
+		}
+
+		/// <summary>
+		/// Zooms the camera
+		/// </summary>
+		/// <param name="action"></param>
+		public void Zoom(eCameraZoomAction action)
+		{
+			if (action == eCameraZoomAction.Stop)
+			{
+				StopZoom();
+				return;
+			}
+
+			SendMoveCommand(s_ZoomActionToCisco[action]);
+			Codec.Log(eSeverity.Informational, "Zooming Far End Camera CallId: {0}, Direction: {1}", m_CallId, action);
+		}
+
+		/// <summary>
+		/// Stops the camera from zooming
+		/// </summary>
+		public void StopZoom()
+		{
+			StopMove();
+		}
+
+		private void SendMoveCommand(string command)
+		{
+			Codec.SendCommand("xCommand Call FarEndControl Camera Move CallId: {0} Value: {1}", m_CallId, command);
+		}
+
+		private void StopMove()
+		{
 			Codec.SendCommand("xCommand Call FarEndControl Camera Stop CallId: {0}", m_CallId);
 			Codec.Log(eSeverity.Informational, "Stop Moving Far End Camera CallId: {0}", m_CallId);
+		}
+
+		#endregion
+
+		#region Console
+
+		/// <summary>
+		/// Gets the child console commands.
+		/// </summary>
+		/// <returns></returns>
+		public override IEnumerable<IConsoleCommand> GetConsoleCommands()
+		{
+			foreach (IConsoleCommand command in GetBaseConsoleCommands())
+				yield return command;
+
+			yield return new EnumConsoleCommand<eCameraZoomAction>("Zoom", e => Zoom(e));
+			yield return new ConsoleCommand("StopZoom", "Stops moving the camera", () => StopZoom());
+		}
+
+		private IEnumerable<IConsoleCommand> GetBaseConsoleCommands()
+		{
+			return base.GetConsoleCommands();
 		}
 
 		#endregion