+++ ICD.Connect.Conferencing.Mock/MockConferencingDeviceSettings.cs
@@ -1,16 +1,10 @@
-﻿using System;
-using ICD.Connect.Devices;
+﻿using ICD.Connect.Devices;
 using ICD.Connect.Settings.Attributes;
 
 namespace ICD.Connect.Conferencing.Mock
 {
-	[KrangSettings(FACTORY_NAME)]
+	[KrangSettings("MockConferencingDevice", typeof(MockConferencingDevice))]
 	public sealed class MockConferencingDeviceSettings : AbstractDeviceSettings, IMockConferencingDeviceSettings
 	{
-		private const string FACTORY_NAME = "MockConferencingDevice";
-
-		public override string FactoryName { get { return FACTORY_NAME; } }
-
-		public override Type OriginatorType { get { return typeof(MockConferencingDevice); } }
 	}
 }