+++ ICD.Connect.Conferencing.Polycom/Devices/Codec/Components/Content/ContentComponent.cs
@@ -8,8 +8,8 @@ using ICD.Connect.API.Nodes;
 
 namespace ICD.Connect.Conferencing.Polycom.Devices.Codec.Components.Content
 {
-	public sealed class ContentComponent : AbstractPolycomComponent
-	{
+	public sealed class ContentComponent : AbstractPolycomComponent, IFeedBackComponent
+    {
 		/// <summary>
 		/// Raised when the active content video source changes.
 		/// </summary>
@@ -70,9 +70,19 @@ namespace ICD.Connect.Conferencing.Polycom.Devices.Codec.Components.Content
 		{
 			base.Initialize();
 
-			Codec.EnqueueCommand("vcbutton register");
+			InitializeFeedBack();
+
 			Codec.EnqueueCommand("vcbutton get");
 			Codec.EnqueueCommand("vcbutton source get");
+		}
+
+
+        /// <summary>
+        /// Called to initialize the feedbacks.
+        /// </summary>
+        public void InitializeFeedBack()
+		{
+			Codec.EnqueueCommand("vcbutton register");
 		}
 
 		#region Methods