+++ ICD.Connect.Conferencing.Polycom/Devices/Codec/Components/Mute/MuteComponent.cs
@@ -9,8 +9,8 @@ using ICD.Connect.API.Nodes;
 
 namespace ICD.Connect.Conferencing.Polycom.Devices.Codec.Components.Mute
 {
-	public sealed class MuteComponent : AbstractPolycomComponent
-	{
+	public sealed class MuteComponent : AbstractPolycomComponent, IFeedBackComponent
+    {
 		private const string MUTE_REGEX = @"mute (?'near'near|far) (?'on'on|off)";
 		private const string VIDEO_MUTE_REGEX = @"videomute near (?'on'on|off)";
 
@@ -130,12 +130,20 @@ namespace ICD.Connect.Conferencing.Polycom.Devices.Codec.Components.Mute
 		{
 			base.Initialize();
 
-			Codec.EnqueueCommand("mute register");
-			Codec.EnqueueCommand("notify mutestatus");
+			InitializeFeedBack();
 
 			Codec.EnqueueCommand("mute near get");
 			Codec.EnqueueCommand("mute far get");
 			Codec.EnqueueCommand("videomute near get");
+		}
+
+		/// <summary>
+		/// Called to initialize the feedbacks.
+		/// </summary>
+		public void InitializeFeedBack()
+		{
+			Codec.EnqueueCommand("mute register");
+			Codec.EnqueueCommand("notify mutestatus");
 		}
 
 		/// <summary>